refactor: Implement review feedback.
